requirements.txt
docker==3.5.1
influxdb==5.2.2
PyYAML==5.1.2